feat(DeploymentManager): manipulate instance group count per pcf tile
Made the bosh deployment manager abstract and implemented one for pcf and one for
the use in other environments. Only the node count will be manipulated in deployment
manifest. Plugins are todo.
